Adding Basic Support for code mods!
This commit reworks the project to be split into two projects - A launcher and a class library for the game. This allows people to more easily add a reference to the game files dll when making code mods.
It also adds a new GameMod class that code mods can extend from to receive various game events, and implemented some events throughout the game, which are controlled by a new mod manager.
And I also set up the process for loading DLLs in the mod loader.