Snap the camera to the nearest 90 degrees if we're close enough
While playing the game I was bothered by the fact that the camera can get misaligned to where it's just a few degrees shy of straight forward. This makes it annoying to line up some jumps. So now, when you get within 18 degrees of one of the four 90 degree directions, you'll snap straight ahead. I played around with the numbers, and this amount of sensitivity feels best. If you think the number should be lower or higher, let me know. Here is a demonstration of the change: https://www.youtube.com/watch?v=sUO0gDmANLQ