Updating Player Class to be more extendable and customizeable. Also allowing mods to more easily replace base game entities with custom ones
This change upgrades the player class in two ways. First, it goes through and makes all the functions and properties public and virtual. this should make it easier for modders to edit properties on the player, and also extend from it to make custom player classes. It also splits the main properties into a default property and a normal property, so we will always have a reference to what the default value is for that property.
It also adds a new status effect system, so mods can more easily apply permanent or temporary status effects to a player, such as changing the gravity or speed for a few seconds.
It also changes the priority for how actor factories are managed, so now, mod actor factories are checked first instead of last. This means. mods can effectively replace actor factories with custom ones, so all actors of a type in a map can be easily replaced with custom ones.