Fixing Localization issues
I noticed and fixed two issues with the localization setup. First it seemed that the Languages wasn't getting properly loaded from the save file. I believe this was because it was just a variable, not a property. Changing it to a property fixed it for me.
Second, I saw it was throwing an error if it did try to load a save file with the language other than English, because Language.Current.Use gets called before the save file is loaded during Assets.Load, so it just uses the default. It needs to get called after the save file is loaded instead of as well, to get around this error.