Mod Menu, Enabling/Disabling Mods and Mod Save Data
This change does three major things.
First, it implements a new mod viewer menu that allows you to view metadata about all of your mods. This is very much a first iteration, and will likely be iterated on later, but it's functional for now.
Second, as part of the mod menu, this adds the ability to Enable and Disable mods, meaning mods that are disabled do not get their events run, they do not have access to Asset Replacement, and skins and maps won't show in their respective menus.
Third, this adds a basic path for modders to save and load their own save data. This is done through 3 passthrough functions on the GameMod class. This may be iterated on later to improve security if that becomes an issue.