Implementing ability to play music and sound effects from wav files
This is an experimental feature that allows modders to play audio from sound effects. It will not be as powerful or fully supported as using fmod files, but it may be useful for smaller or simpler sound effects. This also includes experimental support for using wavs in maps, although this is support may change in the future since it isn't as smart as fmod music.