Changing to no longer throw exception when a mod can't be loaded.
Previously, if a mod couldn't be loaded, it would throw an exception which would crash the game. Now it will just throw an error to the log.
It also now should properly enforce that mod's have unique mod ids, so now if there are two mods with the same id, the second one won't be loaded.