Adding Debug Menu
Adding a basic debug menu that gives some basic debug actions.
This includes loading a map within the current mod, teleporting to a given checkpoint, refilling double dash, and entering a new debug fly mode.
I set this up so it has to be manually enabled through the fuji options so people only use it if they know what they're doing, and also to discourage using it to cheat.