Setting up mod settings menu and ability to autogenerate settings menu from a Settings object.
This commit sets up the core functionality of the settings menu. It updates the menu to use the simpler menu style to work better with the current layout and have more room on screen for settings. It also adds a new GameModSettings file that can be extended by mods. Properties within this object will be be saved to the save file, and will be able to be changed through an autogenerated settings menu based on the provided settings. Currently it supports bools (with Toggles), ints (with sliders), and enums (with multiselects). Settings can be further defined using some new attributes on those properties. The interface for working with this was loosely based off of the Everest Mod Loader, although the implementation is mostly custom.
I also made it easier to grab localization strings for a specific mod, instead of going through the mod asset dictionary.
I also updated the menu system so it will automatically try to localize labels and things for each menu item. There's an optional bool on all the objects to prevent it from localizing for cases where that's useful. I updated the existing menus to account for this.