Fixing localization issues, and reorganizing and documenting some menu code
There were a couple issues that are being addressed in this change. First, I'm reworking how menu item labels and descriptions work. The previous version would not let  you manually override the description, but these changes make that possible again. I also reworked the label so now the LocString and the label are stored as two separate properties, which makes it easier to grab the description from the loc strirng, while also letting us still have control over the final label if we need it.
I also fixed an issue where the mod settings weren't getting localized properly. They were previously watching for the <MISSING> key to determine if the localization worked or not, but since that got removed, it broke those menu items. I fixed by adding a new TryGetModString function which returns whether the mod string exists and its value at the same time.