Adding Hook to sledge to fix issue with some faces disappearing.
There was a bug where sometimes faces were randomly disappearing. I tracked this down to the function in Sledge's Solid class that computes the vertices. That function would incorrectly setup the vertices if there were two faces that had the same Normal, or very similar normals.
I don't know if my solution is perfect, but it seems like I was able to improve it by changing it to check the distance of the plane as well. After my change the examples I had seem to be fixed at least.