Changing so mod loading and unloading happens during full reload instead of when it's toggled
Previously, it was set up so OnModLoaded and OnModUnloaded happened immediately when a mod was toggled. This could create issues because onModLoaded could be called before its assets were loaded, and OnModUnloaded could be called without fully cleaning up its hooks.
So for now at least, I've changed it so OnModLoaded and OnModUnloaded only happen during the full reload that normally happens when you back all the way out of the menu.
I also fixed an issue where the Full reload could get missed if the player pressed Save and Quit before the reload happens. Not it should trigger a reload if they enable/disable a mod and then save and quit without backing out of the menu.