Setting up loading zone functionality
This change adds a new Loading Zone type of solid that allows mappers to more easily teleport between maps and checkpoints. It works similarly to cassettes, but allows you to use a brush to define the trigger space instead, and does not force you into a submap or care if you've entered it before. It also lets you specify which checkpoint you want to start at in the map. Technically, you can even move to a different checkpoint in the same map if desired.