feat: Make error handling smarter, better.
This changes a few things: Error handling is now global, applying to every scene. This allows us to hopefully handle errors smoothly in most places.
Added a separate error message scene. Players are kicked out to this scene when an error occurs.
The old soft crash handler from the World scene is preserved in case of milder crashes.