Reworking DebugActorMenu to be a submenu, fixing warnings
This commit does a few things. The big one is that I changed it to make the Debug Actor Menu be more closely tied to the main debug menu. Previously every panel would get a new window, but it seemed to get a little cluttered, so I sort of reworked it to share the same window and be contextual. I think we may still want to revisit this more later to be a bit smarter in a separate PR.
I also fixed a handful of build warnings that were caused by incorrectly trying to access nullable values.
I also added a destroy actor function, in case that ends up being useful.
And I did a little bit of other minor cleanup and fixes.